NEW CHANGES:
Clock.java: 

adjusted calcTime method, so that it works for combinations of am and pm, and am and am, and pm and pm. Created a calcNetEarnings method, which takes gross pay and applies U.S Taxes. 

TimeClockGUI.java: 

Default time for every day is set to 12am to 12pm. Created the updatePayMethod which calls calcGross and calcNet in Clock.java to update net and gross jLabels. Updated actionListeners for update pay.

NOTE: Saturday is not working due to something in updateCalcHours. Probably has something to do with the size of the array.